requirements.txt
asyncpg==0.31.0
anthropic>=0.40.0
apscheduler==3.10.4
resend==2.0.0
python-dotenv==1.0.0
sentry-sdk==2.0.0
requests>=2.32.0
pytest==8.0.0
pytest-asyncio==0.23.0
